back/pyproject.toml
[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
)/
'''

[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["Benjamin Kane <bene@tapes.co>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.4"
graphene-django = "^2.9.0"
django-filter = "^2.2.0"
djangorestframework = "^3.11.0"
django-cors-headers = "^3.2.1"
confidential = "^2.3.0"
psycopg2 = "^2.8.4"
random-word = "^1.0.4"
graphene-subscriptions = "^1.0.2"
psycopg2-binary = "^2.8.4"

[tool.poetry.dev-dependencies]
pytest = "^5.4.0"
pytest-cov = "^2.8.1"
pytest-django = "^3.8.0"
pytest-xdist = "^1.31.0"
black = {version = "^19.10b0", allow-prereleases = true}
pre-commit = "^2.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
